add weekly request - [ ] Add day of the week - [ ] Add search - [ ] Add splash screen

+++ datamodule/src/main/java/com/example/datamodule/storage/CurrentWeather/WindEntity.kt
@@ -3,7 +3,7 @@ package com.example.datamodule.storage.CurrentWeather
 import com.example.domain.models.CurrentWeather.Wind_Model
 
 data class WindEntity(
-    val speed: Long,
+    val speed: Double,
     val deg: Long,
 ) {
     companion object {}
