+++ domain/src/main/java/com/example/domain/models/CurrentWeather/Wind_Model.kt
@@ -1,6 +1,6 @@
 package com.example.domain.models.CurrentWeather
 
 data class Wind_Model(
-    val speed: Long,
+    val speed: Double,
     val deg: Long,
 )
